Refactored Implosion handling from DemoPets
